requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

azure # MIT License
pbr>=2.0 # Apache-2.0
falcon>=1.0.0 # Apache-2.0
oslo.config>=4.0.0  # Apache-2.0
oslo.concurrency>=3.8.0         # Apache-2.0
oslo.context>=2.14.0  # Apache-2.0
oslo.db>=4.21.1  # Apache-2.0
oslo.i18n>=2.1.0,!=3.15.2  # Apache-2.0
oslo.log>=3.22.0  # Apache-2.0
oslo.middleware>=3.27.0                  # Apache-2.0
oslo.policy>=1.23.0  # Apache-2.0
oslo.serialization>=1.10.0,!=2.19.1      # Apache-2.0
PyMySQL>=0.7.6  # MIT License
keystoneauth1>=2.21.0  # Apache-2.0
keystonemiddleware>=4.12.0  # Apache-2.0
python-glanceclient>=2.7.0  # Apache-2.0
python-neutronclient>=6.3.0  # Apache-2.0
python-novaclient>=9.0.0  # Apache-2.0
openstacksdk==0.9.17  # Apache-2.0
pycryptodome==3.4.7 # BSD 2-clause
